+++ src/scripts/execute-script.ts
@@ -69,10 +69,11 @@ function handleResponse(response: string) {
  * @param originalFilepath The abs filepath to the source filepath, will be used to set the python var `__file__`
  * @param additionalPrint Additional text to be printed to the output once the code has been executed
  */
-function writeDataFile(fileToExecute: string, originalFilepath: string, additionalPrint = "") {
+function writeDataFile(fileToExecute: string, originalFilepath: string, additionalPrint = "", nameVariable = "") {
     let data: any = {};
     data["file"] = fileToExecute;
     data["__file__"] = originalFilepath;
+    data["__name__"] = nameVariable;
     if (additionalPrint) {
         data["additionalPrint"] = additionalPrint;
     }
@@ -174,6 +175,7 @@ export async function execute() {
     if (!vscode.window.activeTextEditor) {
         return;
     }
+    const extConfig = utils.getExtensionConfig();
     const activeDocuemt = vscode.window.activeTextEditor.document;
     const selectedCode = getSelectedTextAsExecutableString();
 
@@ -195,10 +197,11 @@ export async function execute() {
 
     // File an info file telling mb what script to run, etc.
     const additionalPrint = isDebuggingMotionBuilder() ? ">>>" : "";
-    writeDataFile(fileToExecute, activeDocuemt.uri.fsPath, additionalPrint);
+    const nameVar: string | undefined = extConfig.get("execute.name");
+    writeDataFile(fileToExecute, activeDocuemt.uri.fsPath, additionalPrint, nameVar);
 
     // Clear the output channel if enabled in user settings
-    if (utils.getExtensionConfig().get("execute.clearOutput")) {
+    if (extConfig.get("execute.clearOutput")) {
         const outputChannel = getOutputChannel(false);
         if (outputChannel) {
             outputChannel.clear();